requirements.txt
# Django and Core Dependencies
Django==4.2.5
django-environ==0.11.2
Pillow==10.0.0

# GeoDjango Dependencies
psycopg2-binary==2.9.7
GDAL==3.7.1

# API and JSON handling
djangorestframework==3.14.0
django-cors-headers==4.2.0

# Development and Testing
black==23.7.0
flake8==6.0.0
isort==5.12.0
pre-commit==3.3.3
bandit==1.7.5
pylint==3.0.0a7
pylint-django==2.5.3
coverage==7.3.0

# Security
django-csp==3.7
django-ratelimit==4.0.0

# Internationalization
django-modeltranslation==0.18.11

# Admin Interface Enhancements
django-admin-interface==0.26.0

# File handling and storage
django-storages==1.14.0

# Form handling
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Utilities
python-slugify==8.0.1
python-decouple==3.8

# Deployment / WSGI
gunicorn==22.1.0
whitenoise==6.4.0
python-dotenv==1.0.0
